state fix (stable): work around fix have to find where the check_for_collisions function is called from the Menu state

+++ pong/src/physics.rs
@@ -2,6 +2,7 @@ use bevy::{
     prelude::*,
     sprite::collide_aabb::{collide, Collision},
 };
+use crate::states::GameState;
 
 //use crate::scoreboard::Scoreboard;
 use crate::ball::{Ball, ResetBallEvent};
@@ -94,11 +95,16 @@ pub fn check_for_collisions(
     collider_query: Query<(Entity, &Transform), With<Collider>>,
     mut collision_events: EventWriter<CollisionEvent>,
     mut text_query: Query<(&mut Text, &mut Score)>,
-    mut reset_ball_event: EventWriter<ResetBallEvent>
+    mut reset_ball_event: EventWriter<ResetBallEvent>,
+    state: ResMut<State<GameState>>
 ) {
     //
     // Gets the only ball we have and stores the velocity and transform values in variables
     //
+
+    if state.current() == &GameState::Menu {
+        return;
+    }
 
     let (mut ball_velocity, ball_transform) = ball_query.single_mut();
     let ball_size: Vec2 = ball_transform.scale.truncate();
